+++ managers/__init__.py
@@ -1,4 +1,11 @@
 from .permission_manager import PermissionManager
 from .binding_manager import GroupBindingManager
+from .server_manager import ServerProfile, ServerRegistry, build_server_profiles
 
-__all__ = ["PermissionManager", "GroupBindingManager"]
+__all__ = [
+    "PermissionManager",
+    "GroupBindingManager",
+    "ServerProfile",
+    "ServerRegistry",
+    "build_server_profiles",
+]